templates/requirement.txt
mysql-connector==2.2.9
Jinja2==2.10.1